+++ clang-tools-extra/clang-tidy/embedded/EmbeddedTidyModule.cpp
@@ -10,6 +10,7 @@
 #include "../ClangTidyModuleRegistry.h"
 #include "MissingCompoundCheck.h"
 #include "MissingElseCheck.h"
+#include "NoCommaOperatorCheck.h"
 #include "NoEarlyReturnCheck.h"
 #include "NoUnionCheck.h"
 
@@ -23,6 +24,8 @@ public:
         "embedded-missing-compound");
     CheckFactories.registerCheck<MissingElseCheck>(
         "embedded-missing-else");
+    CheckFactories.registerCheck<NoCommaOperatorCheck>(
+        "embedded-no-comma-operator");
     CheckFactories.registerCheck<NoEarlyReturnCheck>(
         "embedded-no-early-return");
     CheckFactories.registerCheck<NoUnionCheck>(